backend_api/requirements.txt
django
djangorestframework
pymysql
mysqlclient
requests
django-cors-headers
pandas